Separate onPostBuild from gatsby-node
